Move response functions to trait
