+++ app/Http/Controllers/Api/AuthController.php
@@ -2,15 +2,17 @@
 
 namespace App\Http\Controllers\API;
 
+use App\Http\Controllers\Controller;
+use App\Traits\ApiResponder;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use App\Http\Controllers\API\BaseController as BaseController;
 use App\Models\User;
 use Validator;
 
 
-class AuthController extends BaseController
+class AuthController extends Controller
 {
+    use ApiResponder;
 
     public function login(Request $request)
     {
