+++ app/Http/Controllers/Api/UserController.php
@@ -4,17 +4,18 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UserRequest;
 use App\Models\User;
+use App\Traits\ApiResponder;
 
 class UserController extends Controller
 {
+    use ApiResponder;
+
     public function update(UserRequest $request, $id)
     {
         $user = User::findOrFail($id);
 
         $user->update($request->validated());
 
-        return response()->json([
-            'user' => $user
-        ], 201);
+        return $this->handleResponse($user, 201);
     }
 }
